refactor auth helpers
