+++ frontend/src/auth.js
@@ -1,0 +1,1 @@
+export { attemptLogin, attemptRegister } from '../public/js/auth.js';
